refactor: extract profile data migrations
