+++ js/profile.js
@@ -1,25 +1,15 @@
 // @ts-check
-// profile.js — Profile CRUD, sex/DOB, location, data migration
+// profile.js — Profile CRUD, sex/DOB, location, and storage facade.
 
 import { state } from './state.js';
-import { MARKER_SCHEMA } from './schema.js';
-import { SPECIALTY_MARKER_DEFS } from './adapters.js';
 import { COUNTRY_LATITUDES, LATITUDE_BANDS } from './constants.js';
 import { callClaudeAPI } from './api.js';
 import { isDebugMode, showConfirmDialog, showNotification } from './utils.js';
 import { encryptedSetItem, encryptedGetItem, encryptedRemoveItem } from './crypto.js';
-import { normalizeLightEnvironmentEveningFields } from './light-env-evening.js';
-import {
-  deleteLabEntryMarker,
-  renameLabEntryMarker,
-  setLabEntryMarker,
-  syncLabEntryInsulinMirror,
-} from './lab-entry.js';
-import { normalizeContextSourceSettings } from './context-source-registry.js';
-import { repairProfileMarkerData } from './profile-marker-migrations.js';
+import { migrateProfileData } from './profile-data-migrations.js';
 import { profileStorageKey } from './profile-storage-key.js';
 
-export { profileStorageKey };
+export { migrateProfileData, profileStorageKey };
 
 const profileDeps = {
   callClaudeAPI,
@@ -304,253 +294,6 @@ function queueProfileSync(profileId, importedData = null) {
   import('./sync.js').then(m => m.onProfileSaved?.(profileId, importedData)).catch(() => {});
 }
 
-/**
- * @param {ProfileData} data
- * @returns {ProfileData}
- */
-export function migrateProfileData(data) {
-  // Migrate sleepCircadian → sleepRest (sleep fields go to sleepRest, circadian items to lightCircadian)
-  if (data.sleepCircadian && !data.sleepRest) {
-    const sc = data.sleepCircadian;
-    if (typeof sc === 'string') {
-      data.sleepRest = sc.trim() ? { duration: null, quality: null, schedule: null, issues: [], note: sc.trim() } : null;
-    } else if (typeof sc === 'object') {
-      const sleepIssues = (sc.issues || []).filter(i => !['blue light blockers', 'morning sunlight'].includes(i));
-      const circadianPractices = (sc.issues || []).filter(i => ['blue light blockers', 'morning sunlight'].includes(i));
-      data.sleepRest = { duration: sc.duration || null, quality: sc.quality || null, schedule: sc.schedule || null, issues: sleepIssues, note: sc.note || '' };
-      if (circadianPractices.length && !data.lightCircadian) {
-        data.lightCircadian = { practices: circadianPractices, timing: null, mealTiming: [], note: '' };
-      }
-    }
-  }
-  delete data.sleepCircadian;
-  // Merge old circadian + sleep strings → sleepRest (very old legacy)
-  if (!data.sleepRest) {
-    const parts = [data.circadian, data.sleep].filter(s => s && s.trim());
-    if (parts.length) data.sleepRest = { duration: null, quality: null, schedule: null, issues: [], note: parts.join('\n\n') };
-  }
-  delete data.circadian;
-  delete data.sleep;
-  // Merge fieldExperts + fieldLens → interpretiveLens
-  if (!data.interpretiveLens) {
-    const parts = [data.fieldExperts, data.fieldLens].filter(s => s && s.trim());
-    if (parts.length) data.interpretiveLens = parts.join('\n\n');
-  }
-  delete data.fieldExperts;
-  delete data.fieldLens;
-  // Migrate string fields → structured objects
-  if (typeof data.diagnoses === 'string') {
-    data.diagnoses = data.diagnoses.trim() ? { conditions: [], note: data.diagnoses.trim(), familyHistory: [] } : null;
-  }
-  // Backfill familyHistory on existing diagnoses objects from before v1.7.
-  // Every read site uses Array.isArray() defensively, so this is purely a
-  // tidiness fix — but keeps `diag.familyHistory.length` valid without
-  // optional chaining everywhere downstream.
-  if (data.diagnoses && typeof data.diagnoses === 'object' && !Array.isArray(data.diagnoses.familyHistory)) {
-    data.diagnoses.familyHistory = [];
-  }
-  if (typeof data.diet === 'string') {
-    data.diet = data.diet.trim() ? { type: null, restrictions: [], pattern: null, note: data.diet.trim() } : null;
-  }
-  if (typeof data.exercise === 'string') {
-    data.exercise = data.exercise.trim() ? { frequency: null, types: [], intensity: null, dailyMovement: null, note: data.exercise.trim() } : null;
-  }
-  if (typeof data.sleepRest === 'string') {
-    data.sleepRest = data.sleepRest.trim() ? { duration: null, quality: null, schedule: null, issues: [], note: data.sleepRest.trim() } : null;
-  }
-  // Migrate old lightCircadian format (had practices/timing) → new format (amLight/daytime/uvExposure/evening/cold/grounding/latitude)
-  if (data.lightCircadian && data.lightCircadian.timing && !data.lightCircadian.amLight) {
-    const old = data.lightCircadian;
-    /** @type {{
-     *   amLight: string | null,
-     *   daytime: string | null,
-     *   uvExposure: string | null,
-     *   evening: string[],
-     *   cold: string | null,
-     *   grounding: string | null,
-     *   latitude: number | null,
-     *   mealTiming: string[],
-     *   note: string
-     * }} */
-    const newLc = { amLight: null, daytime: null, uvExposure: null, evening: [], cold: null, grounding: null, latitude: null, mealTiming: old.mealTiming || [], note: old.note || '' };
-    if (old.practices && old.practices.length) {
-      for (const p of old.practices) {
-        if (p === 'morning sunlight') newLc.amLight = 'morning outdoor (after sunrise)';
-        else if (p === 'blue light blockers') newLc.evening.push('blue blockers after sunset');
-        else if (p === 'no screens before bed') newLc.evening.push('no screens 1-2h before bed');
-        else if (p === 'red light therapy') { if (!newLc.note) newLc.note = p; else newLc.note += ', ' + p; }
-        else if (p === 'UVB exposure') newLc.uvExposure = 'UVB lamp';
-        else if (p === 'light therapy lamp') { if (!newLc.amLight) newLc.amLight = 'light therapy lamp'; }
-        else if (p === 'blackout curtains') { /* moved to sleep environment */ }
-      }
-    }
-    data.lightCircadian = newLc;
-  }
-  // Remove singlePoint from fatty acid custom markers (FA now supports trends)
-  if (data.customMarkers) {
-    for (const def of Object.values(data.customMarkers)) {
-      if (def.singlePoint && def.group === 'Fatty Acids') delete def.singlePoint;
-    }
-  }
-  // Migrate hardcoded specialty markers to customMarkers
-  if (data.entries?.length) {
-    const usedSpecialtyKeys = new Set();
-    for (const entry of data.entries) {
-      for (const key of Object.keys(entry.markers || {})) {
-        if (SPECIALTY_MARKER_DEFS[key]) usedSpecialtyKeys.add(key);
-      }
-    }
-    if (!data.customMarkers) data.customMarkers = {};
-    for (const key of usedSpecialtyKeys) {
-      if (!data.customMarkers[key]) {
-        const def = SPECIALTY_MARKER_DEFS[key];
-        data.customMarkers[key] = {
-          name: def.name, unit: def.unit,
-          refMin: def.refMin, refMax: def.refMax,
-          categoryLabel: def.categoryLabel, icon: def.icon,
-          group: def.group || null
-        };
-      }
-    }
-  }
-  // Backfill group for existing customMarkers missing it
-  if (data.customMarkers) {
-    for (const [key, cm] of Object.entries(data.customMarkers)) {
-      if (cm.group === undefined && SPECIALTY_MARKER_DEFS[key]) {
-        cm.group = SPECIALTY_MARKER_DEFS[key].group || null;
-      }
-    }
-  }
-  repairProfileMarkerData(data);
-  // Fix corrupted FA-prefixed standard markers (bug: _normalizeFattyAcidMarkers rewrote blood work to FA categories)
-  if (data.customMarkers && data.entries?.length) {
-    // Phase 1: relocate markers whose key matches a standard schema marker
-    const _stdLookup = {};
-    for (const [catKey, cat] of Object.entries(MARKER_SCHEMA)) {
-      for (const mk of Object.keys(cat.markers)) _stdLookup[mk] = `${catKey}.${mk}`;
-    }
-    const toDelete = [];
-    for (const fullKey of Object.keys(data.customMarkers)) {
-      const [catKey, markerKey] = fullKey.split('.');
-      if (!markerKey || MARKER_SCHEMA[catKey]) continue;
-      // Don't relocate legitimate specialty markers that happen to share a name with standard ones
-      if (SPECIALTY_MARKER_DEFS[fullKey]) continue;
-      const stdKey = _stdLookup[markerKey];
-      if (!stdKey) continue;
-      for (const entry of data.entries) {
-        renameLabEntryMarker(entry, fullKey, stdKey, { stamp: false });
-      }
-      toDelete.push(fullKey);
-    }
-    for (const key of toDelete) delete data.customMarkers[key];
-    // Phase 2: clean up remaining FA-prefixed markers from entries that also contain standard blood markers.
-    // An entry with both standard markers and FA-prefixed markers = corrupted blood import,
-    // UNLESS the FA marker has a valid customMarker definition (legitimate FA test on same date).
-    const _stdCats = new Set(Object.keys(MARKER_SCHEMA));
-    for (const entry of data.entries) {
-      if (!entry.markers) continue;
-      const keys = Object.keys(entry.markers);
-      const hasStandard = keys.some(k => _stdCats.has(k.split('.')[0]));
-      if (!hasStandard) continue;
-      for (const key of keys) {
-        const catKey = key.split('.')[0];
-        if (!_stdCats.has(catKey) && !SPECIALTY_MARKER_DEFS[key] && (catKey.endsWith('FA') || catKey === 'fattyAcidsTest')) {
-          // Keep markers that have a valid custom marker definition (legitimate FA import)
-          if (data.customMarkers?.[key]) continue;
-          deleteLabEntryMarker(entry, key, { recordTombstone: false, stamp: false });
-        }
-      }
-    }
-    // Phase 3: remove orphaned customMarkers (no values left in any entry)
-    for (const fullKey of Object.keys(data.customMarkers)) {
-      const catKey = fullKey.split('.')[0];
-      if (MARKER_SCHEMA[catKey] || SPECIALTY_MARKER_DEFS[fullKey]) continue;
-      if (!(catKey.endsWith('FA') || catKey === 'fattyAcidsTest')) continue;
-      const hasValues = data.entries.some(e => e.markers?.[fullKey] !== undefined);
-      if (!hasValues) delete data.customMarkers[fullKey];
-    }
-  }
-  // Backfill insulin mirror: sync hormones.insulin ↔ diabetes.insulin_d — v1.6.1
-  if (data.entries) {
-    for (const entry of data.entries) {
-      syncLabEntryInsulinMirror(entry, { stamp: false });
-    }
-  }
-  // Migrate trombocrit/plateletcrit custom markers → hematology.pct — v1.6.1
-  if (data.entries && data.customMarkers) {
-    const pctAliases = Object.keys(data.customMarkers).filter(k =>
-      /tromb|plateletcrit|thrombocrit/i.test(k) && k !== 'hematology.pct'
-    );
-    for (const oldKey of pctAliases) {
-      for (const entry of data.entries) {
-        renameLabEntryMarker(entry, oldKey, 'hematology.pct', { stamp: false });
-      }
-      delete data.customMarkers[oldKey];
-    }
-  }
-  // Migrate hematocrit from fraction (0.45) to percentage (45%) — v1.6.1
-  if (data.entries) {
-    for (const entry of data.entries) {
-      const hct = entry.markers?.['hematology.hematocrit'];
-      if (hct != null && hct < 1) {
-        setLabEntryMarker(entry, 'hematology.hematocrit', parseFloat((hct * 100).toFixed(1)), { stamp: false });
-      }
-    }
-  }
-  // Migrate hematocrit refOverrides from fraction to percentage
-  if (data.refOverrides?.['hematology.hematocrit']) {
-    const ovr = data.refOverrides['hematology.hematocrit'];
-    if (ovr.refMin != null && ovr.refMin < 1) ovr.refMin = parseFloat((ovr.refMin * 100).toFixed(1));
-    if (ovr.refMax != null && ovr.refMax < 1) ovr.refMax = parseFloat((ovr.refMax * 100).toFixed(1));
-    if (ovr.optimalMin != null && ovr.optimalMin < 1) ovr.optimalMin = parseFloat((ovr.optimalMin * 100).toFixed(1));
-    if (ovr.optimalMax != null && ovr.optimalMax < 1) ovr.optimalMax = parseFloat((ovr.optimalMax * 100).toFixed(1));
-  }
-  // Initialize new fields if missing
-  if (data.healthGoals === undefined) data.healthGoals = [];
-  if (data.sleepRest === undefined) data.sleepRest = null;
-  if (data.lightCircadian === undefined) data.lightCircadian = null;
-  if (data.stress === undefined) data.stress = null;
-  if (data.loveLife === undefined) data.loveLife = null;
-  if (data.environment === undefined) data.environment = null;
-  if (data.interpretiveLens === undefined) data.interpretiveLens = '';
-  if (data.contextNotes === undefined) data.contextNotes = '';
-  if (data.customMarkers === undefined) data.customMarkers = {};
-  if (data.menstrualCycle === undefined) data.menstrualCycle = null;
-  if (data.emfAssessment === undefined) data.emfAssessment = null;
-  if (data.emfAssessment && !Array.isArray(data.emfAssessment.assessments)) data.emfAssessment = null;
-  if (data.genetics === undefined) data.genetics = null;
-  if (data.markerNotes === undefined) data.markerNotes = {};
-  if (data.markerValueNotes === undefined) data.markerValueNotes = {};
-  if (data.biologyScoreAI === undefined) data.biologyScoreAI = {};
-  data.contextSourceSettings = normalizeContextSourceSettings(data.contextSourceSettings);
-  if (data.changeHistory === undefined) data.changeHistory = [];
-  if (data.importSnapshots === undefined) data.importSnapshots = [];
-  if (data.biometrics === undefined) data.biometrics = null;
-  // Light lens (v1.7+): sun sessions, light devices, light environment, on-device measurements
-  if (data.sunSessions === undefined) data.sunSessions = [];
-  if (data.deviceSessions === undefined) data.deviceSessions = [];
-  if (data.lightDevices === undefined) data.lightDevices = [];
-  if (data.lightEnvironment === undefined) data.lightEnvironment = null;
-  normalizeLightEnvironmentEveningFields(data.lightEnvironment);
-  if (data.lightMeasurements === undefined) data.lightMeasurements = [];
-  if (data.lightAudits === undefined) data.lightAudits = [];
-  if (data.sunCorrelations === undefined) data.sunCorrelations = null;
-  if (data.lifelightProfile === undefined) data.lifelightProfile = null;
-  if (data.sunDefaults === undefined) data.sunDefaults = null;
-  // Migration — sunDefaults.location → sunDefaults.coords. Earlier demo
-  // imports + a brief window of the v1.6.55 demo upgrade wrote the
-  // location under `.location`, but getSunCoords() (sun.js:2329) reads
-  // `.coords`. Self-heal so the conditions strip + session start dialog
-  // see the location without a re-import.
-  if (data.sunDefaults && data.sunDefaults.location && !data.sunDefaults.coords) {
-    const { lat, lon, label } = data.sunDefaults.location;
-    if (Number.isFinite(lat) && Number.isFinite(lon)) {
-      data.sunDefaults.coords = { lat, lon, source: 'profile-precise', ...(label ? { label } : {}) };
-    }
-    delete data.sunDefaults.location;
-  }
-  return data;
-}
 
 /**
  * @param {string} profileId
