+++ tests/test-unit-import.js
@@ -36,6 +36,7 @@ const reviewRuntimeSrc = read('js/pdf-import-review-runtime.js');
 const progressSrc = read('js/pdf-import-progress.js');
 const labEntrySrc = read('js/lab-entry.js');
 const profileSrc = read('js/profile.js');
+const profileDataMigrationsSrc = read('js/profile-data-migrations.js');
 const importCssSrc = read('css/import.css');
   // ═══════════════════════════════════════
   // 1. normalizeToSI function exists
@@ -151,7 +152,7 @@ const importCssSrc = read('css/import.css');
       && /clearTombstone\(current,\s*['"]importSnapshots['"],\s*snap\.id\)/.test(exportImportSrc)
       && /incomingAt >= existingAt/.test(exportImportSrc));
   assert('Existing profile migration backfills importSnapshots array',
-    /if \(data\.importSnapshots === undefined\) data\.importSnapshots = \[\]/.test(profileSrc));
+    /if \(data\.importSnapshots === undefined\) data\.importSnapshots = \[\]/.test(profileDataMigrationsSrc));
   assert('Import snapshots are tombstone-aware delta array records',
     /importSnapshots:\s*\{[\s\S]{0,180}itemIdFn/.test(read('js/sync-delta-surface-config.js'))
       && /recordTombstone\(state\.importedData,\s*['"]importSnapshots['"],\s*snapId\)/.test(confirmBlock)
